+++ src/controllers/backup.js
@@ -11,7 +11,7 @@ class BackupController {
   }
 
   async createViewAction(request, response) {
-    let user = request.session.user;
+    let user = request.user;
     let readOnly = tools.readOnlyMode() && !user.isAdmin;
     if (readOnly) return response.status(403).json({});
 
@@ -153,7 +153,7 @@ class BackupController {
     }
     let downloadRequestName = request.params.name + '-result-download-request-' + Math.floor(Date.now() / 1000);
     // access
-    if (!tools.hasAccess(request.session.user, backup)) {
+    if (!tools.hasAccess(request.user, backup)) {
       return response.status(403).json({});
     }
     // create download request for result
@@ -213,7 +213,7 @@ class BackupController {
         tools.debug('backup log download : ' + (logResult ? logResult.substring(0, 120) : '') + '...');
       }
       // audit
-      tools.audit(request.session.user.username, 'BackupController', 'DOWNLOAD', request.params.name, 'Backup');
+      tools.audit(request.user.username, 'BackupController', 'DOWNLOAD', request.params.name, 'Backup');
 
       return this.twing
         .render('result.html.twig', {
@@ -238,12 +238,12 @@ class BackupController {
     // filter
     let availableBackups = [];
     for (let i in backups) {
-      if (tools.hasAccess(request.session.user, backups[i])) {
+      if (tools.hasAccess(request.user, backups[i])) {
         availableBackups.push(backups[i]);
       }
     }
     // audit
-    tools.audit(request.session.user.username, 'BackupController', 'LIST', '', 'Backup');
+    tools.audit(request.user.username, 'BackupController', 'LIST', '', 'Backup');
 
     response.type('json').send(sanitizer.sanitize(JSON.stringify(availableBackups)));
   }
@@ -259,14 +259,14 @@ class BackupController {
       return response.status(404).json({});
     }
     // access
-    if (!tools.hasAccess(request.session.user, backup)) {
+    if (!tools.hasAccess(request.user, backup)) {
       return response.status(403).json({});
     }
 
     var deleteRequest = await this.kubeService.createDeleteBackupRequest(request.body.name, request.body.backup);
     if (deleteRequest) {
       // audit
-      tools.audit(request.session.user.username, 'BackupController', 'DELETE', request.params.name, 'Backup');
+      tools.audit(request.user.username, 'BackupController', 'DELETE', request.params.name, 'Backup');
     }
     response.send({ status: deleteRequest ? true : false });
   }