+++ src/controllers/home.js
@@ -7,9 +7,9 @@ class HomeController {
   }
 
   async homeView(request, response) {
-    if (!request.session.user) return response.redirect(tools.subPath('/login'));
+    if (!request.user) return response.redirect(tools.subPath('/login'));
 
-    let user = request.session.user;
+    let user = request.user;
     let readOnly = tools.readOnlyMode() && !user.isAdmin;
     let contexts = this.kubeService.getContexts();
     let currentContext = this.kubeService.getCurrentContext();
@@ -44,7 +44,7 @@ class HomeController {
       }
     }
     // audit
-    tools.audit(request.session.user.username, 'HomeController', 'STATUS');
+    tools.audit(request.user.username, 'HomeController', 'STATUS');
     // check ready
     let isReady = false;
     if (deployStatus && deployStatus.status && deployStatus.status.replicas - deployStatus.status.readyReplicas == 0) {